+++ test/liz/main_test.clj
@@ -104,6 +104,12 @@
                                       (print "Hello\n" []))))))
       "defn- results in private declaration without pub modifier"))
 
+(deftest test-noalias
+  (is (= "pub fn readlink(noalias path: [*:0]const u8, noalias buf_ptr: [*]u8, buf_len: usize) usize;\n"
+         (compile-string (binding [*print-meta* true]
+                           (pr-str '(defn ^usize readlink [^:noalias ^"[*:0]const u8" path ^:noalias ^"[*]u8" buf_ptr ^usize buf_len])))))
+      "noalias parameter modifier"))
+
 ;;(defmacro define-test-cases [suite-name cases results]
 ;;  (let [tests (map vector (eval cases) (eval results))]
 ;;    `(do (is (= ~(count cases)