+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [master](https://github.com/dundalek/closh/compare/v0.1.0...master) (unreleased)
 
 - Changed `defn` to automatically export functions as public and added `defn-` for private declarations
+- Add support for `noalias` function parameter modifier
 
 ## 0.1.0 (2020-12-20)